+++ crates/rucc-types/src/types.rs
@@ -94,7 +94,7 @@ pub struct Types {
     void: TypeId,
     boolean: TypeId,
     ints: [TypeId; 13],
-    floats: [TypeId; 3],
+    floats: [TypeId; 9],
 }
 
 impl Default for Types {
@@ -124,7 +124,7 @@ impl Types {
             void: TypeId(Idx::new(0)),
             boolean: TypeId(Idx::new(0)),
             ints: [TypeId(Idx::new(0)); 13],
-            floats: [TypeId(Idx::new(0)); 3],
+            floats: [TypeId(Idx::new(0)); 9],
         };
         types.void = types.intern(Type::new(TypeKind::Void));
         types.boolean = types.intern(Type::new(TypeKind::Bool));